requirements.txt
annotated-types==0.7.0
branca==0.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
folium==0.19.5
geographiclib==2.0
geopy==2.4.1
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pluggy==1.5.0
pydantic==2.10.6
pydantic_core==2.27.2
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
requests==2.32.3
rich==14.0.0
six==1.17.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
xyzservices==2025.1.0
## libraries needed for the ML Part: (installed in my python3.11.13)
scikit-learn
sentence-transformers
pandas
numpy
matplotlib
seaborn
joblib
jupyter
python-dotenv
requests
aiohttp
umap-learn